+++ CHANGELOG.md
@@ -7,6 +7,8 @@
   approximation accuracy
 * [HexaPDF::Importer::copy] for completely copying (including referenced
   indirect objects) a single PDF object (which may be from the same document)
+* [HexaPDF::Layout::Style::Border#draw_on_bounds] for drawing the border on the
+  bounds instead of inside
 
 ### Changed